Data.Astro.CelestialObject

Description

Computations characteristics of selestial objects.

Documentation

angleEquatorial :: EquatorialCoordinates1 -> EquatorialCoordinates1 -> DecimalDegrees

Calculate angle between two celestial objects whose coordinates specified in Equatorial Coordinate System.

angleEcliptic :: EclipticCoordinates -> EclipticCoordinates -> DecimalDegrees

Calculate angle between two celestial objects whose coordinates specified in Ecliptic Coordinate System.