Data.Astro.Effects

Description

Physical effects which influence on accuracy of astronomical calculations.

Documentation

refract :: DecimalDegrees -> Double -> Double -> DecimalDegrees

Calculate the atmospheric refraction angle. It takes the observed altitude (of Horizon Coordinates), temperature in degrees centigrade and barometric pressure in millibars. The average sea level atmospheric pressure is 1013 millibars.

data AstronomyEpoch

Epoch Enumeration. See also Data.Astro.Time.JulianDate module.

Constructors

B1900
Epoch B1900.0

B1950
Epoch B1950.0

J2000
Epoch J2000.0

J2050
Epoch J2050.0

precession1 :: AstronomyEpoch -> EquatorialCoordinates1 -> JulianDate -> EquatorialCoordinates1

Low-precision method to calculate luni-solar precession. It takes Epoch, Equatorial Coordinates those correct at the given epoch, Julian Date of the observation. It returns corrected Equatorial Coordinates.

precession2 :: JulianDate -> EquatorialCoordinates1 -> JulianDate -> EquatorialCoordinates1

Rigorous method to calculate luni-solar precession. It takes julian date at whose the coordinates are correct, Equatorial Coordinates, Julian Date of the observation. It returns corrected Equatorial Coordinates.

nutationLongitude :: JulianDate -> DecimalDegrees

Calculates the nutation on the ecliptic longitude at the given JulianDate

nutationObliquity :: JulianDate -> DecimalDegrees

Calculates the nutation on the obliquity of the ecliptic at the given JulianDate

includeAberration :: EclipticCoordinates -> JulianDate -> DecimalDegrees -> EclipticCoordinates

Includes aberration effect. It takes true Ecliptic Coordinates, the Sun's longitude at the given Julian Day (the third parameter). Returns apparent ecliptic coordinates. The Sun's longitude can be calculated using sunEclipticLongitude1 or sunEclipticLongitude2 of Data.Astro.Sun module.

parallax :: GeographicCoordinates -> Double -> AstronomicalUnits -> JulianDate -> EquatorialCoordinates1 -> EquatorialCoordinates1

Calculate the apparent position of the celestial object (the Sun or a planet). It takes geocraphic coordinates of the observer and height above sea-level of the observer measured in metres, distance from the celestial object to the Earth measured in AU, the Universal Time and geocentric equatorial coordinates. It returns adjusted equatorial coordinates.