Data.Astro.Effects.Parallax

Description

Calculation effects of geocentric parallax.

Documentation

parallaxQuantities :: DecimalDegrees -> Double -> (Double, Double)

It takes latitude of the observer and height above sea-level of the observer measured in metres Returns palallax quantities (p*(sin phi'), p*(cos phi')), where phi' is the geocentric latitude and p is the distance of the obserbve from the centre of the Earth.

parallax :: GeographicCoordinates -> Double -> AstronomicalUnits -> JulianDate -> EquatorialCoordinates1 -> EquatorialCoordinates1

Calculate the apparent position of the celestial object (the Sun or a planet). It takes geocraphic coordinates of the observer and height above sea-level of the observer measured in metres, distance from the celestial object to the Earth measured in AU, the Universal Time and geocentric equatorial coordinates. It returns adjusted equatorial coordinates.