Data.Astro.Time

Description

Root Time module

Documentation

utToLST :: DecimalDegrees -> JulianDate -> LocalSiderealTime

Universal Time to Local Sidereal Time. It takes longitude in decimal degrees and local civil time

lctToLST :: DecimalDegrees -> LocalCivilTime -> LocalSiderealTime

Local Civil Time to Local Sidereal Time. It takes longitude in decimal degrees and local civil time

lstToLCT :: DecimalDegrees -> LocalCivilDate -> LocalSiderealTime -> LocalCivilTime

Local Sidereal Time to Local Civil Time. It takes longitude in decimal degrees, local civil date and local sidereal time