Data.Astro.Utils

Description

Utility functions.

Documentation

fromFixed :: (Fractional a, HasResolution b) => Fixed b -> a

Convert From Fixed to Fractional

trunc :: RealFrac a => a -> a

return the integral part of a number almost the same as truncate but result type is Real

fraction :: (RealFrac a, Num b) => a -> (b, a)

Almost the same the properFraction function but result type

reduceToZeroRange :: RealFrac a => a -> a -> a

Reduce to range from 0 to n

toRadians :: Floating a => a -> a

Convert from degrees to radians

fromRadians :: Floating a => a -> a

Convert from radians to degrees

roundToN :: RealFrac a => Int -> a -> a

Round to a specified number of digits

tropicalYearLen :: Double

Length of a tropical year in days