Data.Astro.CelestialObject

Description

Computations characteristics of selestial objects.

Documentation

data RiseSet a

Some Info of Rise and Set of a celestial object

Constructors

RiseSet a a
Some Info of Rise and Set of the celestial object

Circumpolar
The celestial object is always above the horizon

NeverRises
The celestial object is always below the horizon

type RSInfo a = (a, DecimalDegrees)

Rise or Set time and azimuth

type RiseSetLST = RiseSet (RSInfo LocalSiderealTime)

LST (Local Sidereal Time) and Azimuth of Rise and Set

type RiseSetLCT = RiseSet (RSInfo LocalCivilTime)

Local Civil Time and Azimuth of Rise and Set

angleEquatorial :: EquatorialCoordinates1 -> EquatorialCoordinates1 -> DecimalDegrees

Calculate angle between two celestial objects whose coordinates specified in Equatorial Coordinate System.

angleEcliptic :: EclipticCoordinates -> EclipticCoordinates -> DecimalDegrees

Calculate angle between two celestial objects whose coordinates specified in Ecliptic Coordinate System.

riseAndSet :: EquatorialCoordinates1 -> DecimalDegrees -> DecimalDegrees -> RiseSetLST

Calculate rise and set local sidereal time of a celestial object. It takes the equatorial coordinates of the celestial object, vertical shift and the latitude of the observation. To calculate vertical shift for stars use function refract from Data.Astro.Effects. In most cases you can assume that vertical shift equals 0.566569 (34 arcmins ~ 'refract (DD 0) 12 1012').

toRiseSetLCT :: DecimalDegrees -> LocalCivilDate -> RiseSetLST -> RiseSetLCT

Converts Rise and Set in Local Sidereal Time to Rise and Set in Local Civil Time.