Data.Astro.Moon

Description

Calculation characteristics of the Moon.

Documentation

moonPosition1 :: MoonDetails -> JulianDate -> EquatorialCoordinates1

Calculate Equatorial Coordinates of the Moon with the given MoonDetails and at the given JulianDate. It is recommended to use j2010MoonDetails as a first parameter.

moonDistance1 :: MoonDetails -> JulianDate -> MoonDistanceUnits

Calculates the Moon's Distance at the given julian date. Returns distance to the Moon moonDistance1 :: JulianDate -> MoonDistanceUnits you can use mduToKm (defined in Data.Astro.Moon.MoonDetails) to convert result to kilometers

moonAngularSize :: MoonDistanceUnits -> DecimalDegrees

Calculate the Moon's angular size at the given distance.

moonHorizontalParallax :: MoonDistanceUnits -> DecimalDegrees

Calculates the Moon's horizontal parallax at the given distance.

moonPhase :: MoonDetails -> JulianDate -> Double

Calculates the Moon's phase (the area of the visible segment expressed as a fraction of the whole disk) at the given universal time.

moonBrightLimbPositionAngle :: EquatorialCoordinates1 -> EquatorialCoordinates1 -> DecimalDegrees

Calculate the Moon's position-angle of the bright limb. It takes the Moon's coordinates and the Sun's coordinates. Position-angle is the angle of the midpoint of the illuminated limb measured eastwards from the north point of the disk.