Data.Astro.Sun

Description

Calculation characteristics of the Sun.

Terms

perihelion - minimal distance from the Sun to the planet

aphelion - maximal distance from the Sun to the planet

perigee - minimal distance from the Sun to the Earth

apogee - maximal distance from the Sun to the Earth

Documentation

data SunDetails

Details of the Sun's apparent orbit at the given epoch

Constructors

SunDetails

Fields

sdEpoch :: JulianDate

Epoch

sdEpsilon :: DecimalDegrees

Ecliptic longitude at the Epoch

sdOmega :: DecimalDegrees

Ecliptic longitude of perigee at the Epoch

sdE :: Double

Eccentricity of the orbit at the Epoch

data RiseSet a

Some Info of Rise and Set of a celestial object

Constructors

RiseSet a a
Some Info of Rise and Set of the celestial object

Circumpolar
The celestial object is always above the horizon

NeverRises
The celestial object is always below the horizon

sunDetails :: JulianDate -> SunDetails

Calculate SunDetails for the given JulianDate.

j2010SunDetails :: SunDetails

SunDetails at the Sun's reference Epoch J2010.0

sunMeanAnomaly2 :: SunDetails -> DecimalDegrees

Calculate mean anomaly using the second 'more accurate' method

sunEclipticLongitude1 :: SunDetails -> JulianDate -> DecimalDegrees

Calculate the ecliptic longitude of the Sun with the given SunDetails at the given JulianDate

sunEclipticLongitude2 :: SunDetails -> DecimalDegrees

Calculate the ecliptic longitude of the Sun

sunPosition1 :: SunDetails -> JulianDate -> EquatorialCoordinates1

Calculate Equatorial Coordinates of the Sun with the given SunDetails at the given JulianDate. It is recommended to use j2010SunDetails as a first parameter.

sunPosition2 :: JulianDate -> EquatorialCoordinates1

More accurate method to calculate position of the Sun

sunDistance :: JulianDate -> Double

Calculate Sun-Earth distance.

sunAngularSize :: JulianDate -> DecimalDegrees

Calculate the Sun's angular size (i.e. its angular diameter).

sunRiseAndSet :: GeographicCoordinates -> DecimalDegrees -> LocalCivilDate -> RiseSetMB

Calculatesthe Sun's rise and set It takes coordinates of the observer, local civil date, vertical shift (good value is 0.833333). It returns Nothing if fails to calculate rise and/or set. It should be accurate to within a minute of time.

equationOfTime :: JulianDate -> DecimalHours

Calculates discrepancy between the mean solar time and real solar time at the given date.

solarElongation :: EquatorialCoordinates1 -> JulianDate -> DecimalDegrees

Calculates the angle between the lines of sight to the Sun and to a celestial object specified by the given coordinates at the given Universal Time.