Data.Astro.Time.JulianDate

Description

Julian date is the continuous count of days since noon on January 1, 4713 BC, the beginning of the Julian Period.

Documentation

newtype JulianDate

A number of days since noon of 1 January 4713 BC

Constructors

JD TimeBaseType

data LocalCivilTime

Represents Local Civil Time

Constructors

LCT

Fields

lctTimeZone :: DecimalHours

lctUniversalTime :: JulianDate

data LocalCivilDate

Local Civil Date, used for time conversions when base date is needed

Constructors

LCD

Fields

lcdTimeZone :: DecimalHours

lcdDate :: JulianDate

type TimeBaseType = Double

numberOfDays :: JulianDate -> JulianDate -> TimeBaseType

Return number of days since the first argument till the second one

numberOfYears :: JulianDate -> JulianDate -> TimeBaseType

Return number of years since the first argument till the second one

numberOfCenturies :: JulianDate -> JulianDate -> TimeBaseType

Return number of centuries since the first argument till the second one

addHours :: DecimalHours -> JulianDate -> JulianDate

add Decimal Hours

fromYMD :: Integer -> Int -> Int -> JulianDate

Create Julian Date. It takes year, month [1..12], Day [1..31].

fromYMDHMS :: Integer -> Int -> Int -> Int -> Int -> TimeBaseType -> JulianDate

Create Julian Date. It takes year, month [1..12], Day [1..31], hours, minutes, seconds.

toYMDHMS :: JulianDate -> (Integer, Int, Int, Int, Int, TimeBaseType)

It returns year, month [1..12], Day [1..31], hours, minutes, seconds.

dayOfWeek :: JulianDate -> Int

Get Day of the Week 0 is for Sunday, 1 for manday and 6 for Saturday

splitToDayAndTime :: JulianDate -> (JulianDate, JulianDate)

Extract Day and Time parts of Date

lctFromYMDHMS :: DecimalHours -> Integer -> Int -> Int -> Int -> Int -> TimeBaseType -> LocalCivilTime

Create LocalCivilTime from tize zone, local year, local month, local day, local hours, local minutes and local secunds.

lctToYMDHMS :: LocalCivilTime -> (Integer, Int, Int, Int, Int, TimeBaseType)

Get from LocalCivilTime local year, local month, local day, local hours, local minutes and local secunds.

lcdFromYMD :: DecimalHours -> Integer -> Int -> Int -> LocalCivilDate