Data.Astro.Time.Sidereal

Description

According to the Sidereal Clock any observed star returns to the same position in the sky every 24 hours.

Each sidereal day is shorter than the solar day, 24 hours of sidereal time corresponding to 23:56:04.0916 of solar time.

Documentation

data GreenwichSiderealTime

Greenwich Sidereal Time GST can be in range [-12h, 36h] carrying out a day correction

data LocalSiderealTime

Local Sidereal Time

dhToGST :: DecimalHours -> GreenwichSiderealTime

Convert Decimal Hours to Greenwich Sidereal Time

dhToLST :: DecimalHours -> LocalSiderealTime

Convert Decimal Hours to Local Sidereal Time

gstToDH :: GreenwichSiderealTime -> DecimalHours

Convert Greenwich Sidereal Time to Decimal Hours

lstToDH :: LocalSiderealTime -> DecimalHours

Convert Local Sidereal Time to Decimal Hours

hmsToGST :: Int -> Int -> TimeBaseType -> GreenwichSiderealTime

Comvert Hours, Minutes, Seconds to Greenwich Sidereal Time

hmsToLST :: Int -> Int -> TimeBaseType -> LocalSiderealTime

Comvert Hours, Minutes, Seconds to Local Sidereal Time

utToGST :: JulianDate -> GreenwichSiderealTime

Convert from Universal Time (UT) to Greenwich Sidereal Time (GST)

gstToUT :: JulianDate -> GreenwichSiderealTime -> JulianDate

Convert from Greenwich Sidereal Time (GST) to Universal Time (UT). It takes GST and Greenwich Date, returns JulianDate. Because the sidereal day is shorter than the solar day (see comment to the module). In case of such ambiguity the early time will be returned. You can easily check the ambiguity: if time is equal or less 00:03:56 you can get the second time by adding 23:56:04

gstToLST :: DecimalDegrees -> GreenwichSiderealTime -> LocalSiderealTime

Convert Greenwich Sidereal Time to Local Sidereal Time. It takes GST and longitude in decimal degrees

lstToGST :: DecimalDegrees -> LocalSiderealTime -> GreenwichSiderealTime

Convert Local Sidereal Time to Greenwich Sidereal Time It takes LST and longitude in decimal degrees

lstToGSTwDC :: DecimalDegrees -> LocalSiderealTime -> GreenwichSiderealTime

Convert Local Sidereal Time to Greenwich Sidereal Time with Day Correction. It takes LST and longitude in decimal degrees