Data.Astro.Types

Description

Common Types are usfull across all subsystems like Time and Coordinate

Documentation

newtype DecimalDegrees

Constructors

DD Double

newtype DecimalHours

Constructors

DH Double

data GeographicCoordinates

Geographic Coordinates

Constructors

GeoC

Fields

geoLatitude :: DecimalDegrees

geoLongitude :: DecimalDegrees

newtype AstronomicalUnits

Astronomical Units, 1AU = 1.4960×1011 m (originally, the average distance of Earth's aphelion and perihelion).

Constructors

AU Double

lightTravelTime :: AstronomicalUnits -> DecimalHours

Light travel time of the distance in Astronomical Units

toDecimalHours :: DecimalDegrees -> DecimalHours

Convert decimal degrees to decimal hours

fromDecimalHours :: DecimalHours -> DecimalDegrees

Convert decimal hours to decimal degrees

toRadians :: DecimalDegrees -> Double

Convert from DecimalDegrees to Radians

fromRadians :: Double -> DecimalDegrees

Convert from Radians to DecimalDegrees

fromDMS :: RealFrac a => Int -> Int -> a -> DecimalDegrees

Convert Degrees, Minutes, Seconds to DecimalDegrees

toDMS :: (Integral t, Integral t1) => DecimalDegrees -> (t, t1, Double)

Convert DecimalDegrees to Degrees, Minutes, Seconds

fromHMS :: RealFrac a => Int -> Int -> a -> DecimalHours

Comvert Hours, Minutes, Seconds to DecimalHours

toHMS :: (Integral t, Integral t1) => DecimalHours -> (t, t1, Double)

Convert DecimalDegrees to Degrees, Minutes, Seconds